LayoutTests/fast/viewport/ios/adjust-fixed-width-and-initial-scale-to-avoid-excessive-zooming-expected.txt

This test verifies that when the `IgnoreMetaViewport` policy is enabled, a meta viewport that specifies an explicit width (with a large initial scale) doesn't cause us to zoom in excessively.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS visualViewport.scale < 2 is true
PASS successfullyParsed is true

TEST COMPLETE
